Semicolon sanitizier
