+++ drools-model/drools-mvel-compiler/src/main/java/org/drools/mvelcompiler/MvelCompiler.java
@@ -40,20 +40,4 @@ public class MvelCompiler {
 
         return new ParsingResult(statements).addModifyProperties(modifiedProperties);
     }
-
-    // TODO this has to be properly parsed
-    public static String sanitizeMvelScript(String mvelScript) {
-        String[] split = mvelScript.split("\n");
-
-        StringBuilder builder = new StringBuilder();
-        builder.append("{");
-        for(String s : split) {
-            builder.append(s);
-            if(!s.trim().endsWith(";")) {
-                builder.append(";");
-            }
-        }
-        builder.append("}");
-        return builder.toString();
-    }
 }
