+++ drools-model/drools-mvel-compiler/src/main/java/org/drools/mvelcompiler/SemicolonSanitizer.java
@@ -1,0 +1,20 @@
+package org.drools.mvelcompiler;
+
+// TODO this class has to be deleted
+public class SemicolonSanitizer {
+
+    public static String sanitizeMvelScript(String mvelScript) {
+        String[] split = mvelScript.split("\n");
+
+        StringBuilder builder = new StringBuilder();
+        builder.append("{");
+        for (String s : split) {
+            builder.append(s);
+            if (!s.trim().endsWith(";")) {
+                builder.append(";");
+            }
+        }
+        builder.append("}");
+        return builder.toString();
+    }
+}
